test: rename directory
